New style nav
I added some styling to the navigation. It looks great but to make ot work I had to shorten the title of the Things you need to know page. And we might not want to do that so feel free to reject this commit! The mail to link works - if I click it on my mac it automatically opens a draft email.